+++ src/ui/react/fancy-toolbar/app.tsx
@@ -3,7 +3,7 @@ import { ErrorBoundary } from "../weg/components/Error/index.tsx";
 import { ErrorFallback } from "./components/Error/index.tsx";
 import { FancyToolbar } from "./modules/main/Toolbar.tsx";
 import { useSignalEffect } from "@preact/signals";
-import { $lastFocusedOnMonitor } from "./modules/shared/state/windows.ts";
+import { $isSomeFullscreenOnMonitor } from "./modules/shared/state/windows.ts";
 import { $initialPositionSet } from "./modules/shared/state/mod.ts";
 import { useEffect } from "preact/hooks";
 
@@ -13,7 +13,7 @@ export function App() {
   }, []);
 
   useSignalEffect(() => {
-    const fullscreened = !!$lastFocusedOnMonitor.value?.isFullscreened;
+    const fullscreened = $isSomeFullscreenOnMonitor.value;
     if (fullscreened) {
       Widget.self.window.setAlwaysOnTop(false);
       Widget.self.window.setAlwaysOnBottom(true);